requirements.txt
asgiref==3.11.1
Django==6.0.2
gunicorn==25.0.3
joblib==1.5.3
numpy==2.4.2
packaging==26.0
psycopg2-binary==2.9.11
python-magic==0.4.27
scikit-learn==1.8.0
scipy==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
